Add files via upload
Overall Review,

The template represents 3-tier Architect. we have separated 2 public subnets for ec2 instance and a private subnet for RDS instances and also 2 public subnets assigned for ALB in different AVzone. 
We have created the listener for 80 port and assigned the target group under the 2 ec2 instances with 80 protocol.

The following steps guide you about this JSON template.
	Step1 - Parameters
		1.AVzone 		-We have created 3 Aavilabiltyzone for single region.
		2.VPC 			-You can assign VPC CIDR value manually here, as per your environment growth. 
		3.EC2subnet 	-You can setup the subnet CIDR value manually based on your VPC and ,Here we have poited 2subnet for EC2 public subnet and one for RDS private subnet.
		4.ALBsubnet 	-You can setup the subnet CIDR value manually based on your VPC and ,Here we have poited 2subnet for ALB public subnet.
		5.NATsubnet 	
		6.ELBIPaddress	
		7.CertificateArn
	Step2 - Resources